+++ demo/app/demo-clustering-custom-render/index.js
@@ -7,6 +7,7 @@ import faker from 'faker';
 import Timeline from 'react-calendar-timeline';
 import timelineData from '../fake-timeline-data';
 import clusterItemRenderer from './ItemRenderCluster';
+import ZoomRenderer from './ZoomRenderer';
 
 const minTime = moment('2022-MAY-12')
   .add(-12, 'hours')
@@ -26,6 +27,83 @@ const keys = {
   itemTimeStartKey: 'start',
   itemTimeEndKey: 'end',
 };
+
+export const zoomStepTimeWindows = [
+  {
+    timespanInMS: 172800000, // 48hrs
+    hour: 4,
+    minute: 1,
+    second: 1,
+    day: 1,
+    month: 1,
+    year: 1,
+  }, {
+    timespanInMS: 115200000, // 32hrs
+    hour: 4,
+    minute: 1,
+    second: 1,
+    day: 1,
+    month: 1,
+    year: 1,
+  }, {
+    timespanInMS: 86400000, // 24hrs
+    hour: 4,
+    minute: 1,
+    second: 1,
+    day: 1,
+    month: 1,
+    year: 1,
+  }, {
+    timespanInMS: 57600000, // 16hrs
+    hour: 4,
+    minute: 1,
+    second: 1,
+    day: 1,
+    month: 1,
+    year: 1,
+  }, {
+    timespanInMS: 43200000, // 12hrs
+    hour: 2,
+    minute: 1,
+    second: 1,
+    day: 1,
+    month: 1,
+    year: 1,
+  }, {
+    timespanInMS: 28800000, // 8hrs
+    hour: 1,
+    minute: 30,
+    second: 1,
+    day: 1,
+    month: 1,
+    year: 1,
+  }, {
+    timespanInMS: 14400000, // 4hrs
+    hour: 1,
+    minute: 30,
+    second: 1,
+    day: 1,
+    month: 1,
+    year: 1,
+  }, {
+    timespanInMS: 7200000, // 2hrs
+    hour: 1,
+    minute: 10,
+    second: 1,
+    day: 1,
+    month: 1,
+    year: 1,
+  }, {
+    timespanInMS: 3600000, // 1hrs
+    hour: 1,
+    minute: 5,
+    second: 1,
+    day: 1,
+    month: 1,
+    year: 1,
+  },
+];
+
 
 const SIXTEEN_HOURS_IN_MS = 1000 * 60 * 60 * 16;
 
@@ -149,6 +227,7 @@ export default class App extends Component {
         sidebarWidth={150}
         sidebarContent={<div>Above The Left</div>}
         canMove
+        zoomRenderer={ZoomRenderer}
         canResize="right"
         canSelect
         itemTouchSendsClick={false}
@@ -158,6 +237,7 @@ export default class App extends Component {
           sequencialClusterTinyItemsOnly: true,
           disableClusteringBelowTime: SIXTEEN_HOURS_IN_MS,
         }}
+        timeSteps={zoomStepTimeWindows}
         itemSorted
         itemRendererCluster={clusterItemRenderer}
         itemHeightRatio={0.75}