+++ src/lib/Timeline.js
@@ -6,6 +6,7 @@ import Sidebar from './layout/Sidebar';
 import ScrollElement from './scroll/ScrollElement';
 import MarkerCanvas from './markers/MarkerCanvas';
 import Rows from './rows/Rows';
+import ZoomControl from './zoom';
 
 import windowResizeDetector from '../resize-detector/window';
 
@@ -48,7 +49,7 @@ export default class ReactCalendarTimeline extends Component {
     minZoom: PropTypes.number,
     maxZoom: PropTypes.number,
     zoomThrottle: PropTypes.number,
-
+    zoomControl: PropTypes.bool,
     clickTolerance: PropTypes.number,
 
     canChangeGroup: PropTypes.bool,
@@ -84,7 +85,7 @@ export default class ReactCalendarTimeline extends Component {
     itemRenderer: PropTypes.func,
     itemRendererCluster: PropTypes.func,
     groupRenderer: PropTypes.func,
-
+    zoomRenderer: PropTypes.func,
     clusterSettings: PropTypes.object,
     // className: PropTypes.string,
     style: PropTypes.object,
@@ -104,21 +105,35 @@ export default class ReactCalendarTimeline extends Component {
     headerRef: PropTypes.func,
     scrollRef: PropTypes.func,
 
-    timeSteps: PropTypes.shape({
-      second: PropTypes.number,
-      minute: PropTypes.number,
-      hour: PropTypes.number,
-      day: PropTypes.number,
-      month: PropTypes.number,
-      year: PropTypes.number,
-    }),
+    timeSteps: PropTypes.oneOfType([
+      PropTypes.shape({
+        second: PropTypes.number,
+        minute: PropTypes.number,
+        hour: PropTypes.number,
+        day: PropTypes.number,
+        month: PropTypes.number,
+        year: PropTypes.number,
+      }),
+      PropTypes.arrayOf(
+        PropTypes.shape(
+          {
+            timespanInMS: PropTypes.number,
+            second: PropTypes.number,
+            minute: PropTypes.number,
+            hour: PropTypes.number,
+            day: PropTypes.number,
+            month: PropTypes.number,
+            year: PropTypes.number,
+          },
+        ),
+      ),
+    ]),
 
     defaultTimeStart: PropTypes.object,
     defaultTimeEnd: PropTypes.object,
 
     zoomTimeStart: number,
     zoomTimeEnd: number,
-    zoomThrottle: number,
 
     visibleTimeStart: PropTypes.number,
     visibleTimeEnd: PropTypes.number,
@@ -180,7 +195,7 @@ export default class ReactCalendarTimeline extends Component {
     stickyHeader: true,
     lineHeight: 30,
     itemHeightRatio: 0.65,
-
+    zoomControl: false,
     minZoom: 60 * 60 * 1000, // 1 hour
     maxZoom: 5 * 365.24 * 86400 * 1000, // 5 years
     zoomThrottle: 1,
@@ -262,6 +277,21 @@ export default class ReactCalendarTimeline extends Component {
     getTimelineContext: PropTypes.func,
   }
 
+  getTimeStep = (timeSteps) => {
+    if (Array.isArray(timeSteps) && timeSteps.length > 0) {
+      const { visibleTimeStart, visibleTimeEnd } = this.state;
+      const visibleTime = visibleTimeEnd - visibleTimeStart;
+      let timeStepToReturn = timeSteps[0];
+      for (let i = 1; i < timeSteps.length; i += 1) {
+        if (timeSteps[i].timespanInMS >= visibleTime) {
+          timeStepToReturn = timeSteps[i];
+        }
+      }
+      return timeStepToReturn;
+    }
+    return timeSteps;
+  }
+
   getChildContext() {
     return {
       getTimelineContext: () => this.getTimelineContext(),
@@ -294,9 +324,8 @@ export default class ReactCalendarTimeline extends Component {
     } = this.state;
 
     const { timeSteps } = this.props;
-
     const zoom = visibleTimeEnd - visibleTimeStart;
-    const minUnit = getMinUnit(zoom, width, timeSteps);
+    const minUnit = getMinUnit(zoom, width, this.getTimeStep(timeSteps));
 
     return minUnit;
   }
@@ -311,12 +340,19 @@ export default class ReactCalendarTimeline extends Component {
     let visibleTimeStart = null;
     let visibleTimeEnd = null;
 
+    let initialTimeStart = null;
+    let initialTimeEnd = null;
+
     if (this.props.defaultTimeStart && this.props.defaultTimeEnd) {
       visibleTimeStart = this.props.defaultTimeStart.valueOf();
       visibleTimeEnd = this.props.defaultTimeEnd.valueOf();
+      initialTimeStart = this.props.defaultTimeStart.valueOf();
+      initialTimeEnd = this.props.defaultTimeEnd.valueOf();
     } else if (this.props.visibleTimeStart && this.props.visibleTimeEnd) {
       visibleTimeStart = this.props.visibleTimeStart;
       visibleTimeEnd = this.props.visibleTimeEnd;
+      initialTimeStart = this.props.visibleTimeStart;
+      initialTimeEnd = this.props.visibleTimeEnd;
     } else {
       // throwing an error because neither default or visible time props provided
       throw new Error(
@@ -333,6 +369,10 @@ export default class ReactCalendarTimeline extends Component {
       width: 1000,
       zoomTimeStart: this.props.zoomTimeStart,
       zoomTimeEnd: this.props.zoomTimeEnd,
+      zoomControl: this.props.zoomControl,
+      zoomScalePercent: 0.7,
+      initialTimeStart,
+      initialTimeEnd,
       visibleTimeStart,
       visibleTimeEnd,
       canvasTimeStart,
@@ -540,9 +580,6 @@ export default class ReactCalendarTimeline extends Component {
       this.state.newGroupId,
       props.clusterSettings,
     );
-
-    // this is needed by dragItem since it uses pageY from the drag events
-    // if this was in the context of the scrollElement, this would not be necessary
 
     this.setState({
       width,
@@ -933,6 +970,7 @@ export default class ReactCalendarTimeline extends Component {
       traditionalZoom,
       itemRenderer,
       itemRendererCluster,
+      zoomRenderer,
       keys,
       hideHorizontalLines,
     } = this.props;
@@ -951,7 +989,7 @@ export default class ReactCalendarTimeline extends Component {
 
     const zoom = visibleTimeEnd - visibleTimeStart;
     const canvasWidth = getCanvasWidth(width);
-    const minUnit = getMinUnit(zoom, width, timeSteps);
+    const minUnit = getMinUnit(zoom, width, this.getTimeStep(timeSteps));
 
     const isInteractingWithItem = !!draggingItem || !!resizingItem;
 
@@ -984,6 +1022,23 @@ export default class ReactCalendarTimeline extends Component {
     const outerComponentStyle = {
       height: `${height}px`,
     };
+
+    const zoomControl = () => {
+      if (zoomRenderer) {
+        return zoomRenderer({
+          onZoomIn: ({ value } = {}) => this.changeZoom(value || 1 * this.state.zoomScalePercent),
+          onZoomOut: ({ value } = {}) => this.changeZoom(value || 1 / this.state.zoomScalePercent),
+          onZoomReset: ({ canvasTimeStartZoom, canvasTimeEndZoom } = {}) => this.showPeriod(canvasTimeStartZoom || this.state.initialTimeStart, canvasTimeEndZoom || this.state.initialTimeEnd),
+        });
+      }
+      if (this.props.zoomControl) {
+        return <ZoomControl
+        onZoomIn={() => this.changeZoom(1 * this.state.zoomScalePercent)}
+        onZoomOut={() => this.changeZoom(1 / this.state.zoomScalePercent)}
+        onZoomReset={() => this.showPeriod(this.state.initialTimeStart, this.state.initialTimeEnd)}/>;
+      }
+      return undefined;
+    };
     return (
       <TimelineStateProvider
         visibleTimeStart={visibleTimeStart}
@@ -999,7 +1054,7 @@ export default class ReactCalendarTimeline extends Component {
         <TimelineMarkersProvider>
           <TimelineHeadersProvider
             registerScroll={this.handleHeaderRef}
-            timeSteps={timeSteps}
+            timeSteps={this.getTimeStep(timeSteps)}
             leftSidebarWidth={this.props.sidebarWidth}
             rightSidebarWidth={this.props.rightSidebarWidth}
           >
@@ -1015,6 +1070,7 @@ export default class ReactCalendarTimeline extends Component {
                 ref={el => (this.container = el)}
                 className="react-calendar-timeline"
               >
+                 {zoomControl()}
                 {this.renderHeaders()}
                 <div style={outerComponentStyle} className="rct-outer">
                   {sidebarWidth > 0 ? this.sidebar(height, groupHeights) : null}
@@ -1072,7 +1128,7 @@ export default class ReactCalendarTimeline extends Component {
                       {hideHorizontalLines ? null : <Columns
                         lineCount={_length(groups)}
                         minUnit={minUnit}
-                        timeSteps={timeSteps}
+                        timeSteps={this.getTimeStep(timeSteps)}
                         verticalLineClassNamesForTime={this.props.verticalLineClassNamesForTime}
                         canvasTimeStart={canvasTimeStart}
                         canvasTimeEnd={canvasTimeEnd}